# Approximation of latToY function