@TheMarex 2016-05-05
Incredible work. This will also speed up other algorithms like Douglas-Peucker. Can't wait to do a full-stack benchmark with this.