@emiltin 2016-05-05
excellent. but what happens outside of the -70 to 70 range?