@danpat 2016-05-05
@emiltin well, things start to go a little haywire after about +/- 70.5:

![screenshot from 2016-05-05 13-16-18](https://cloud.githubusercontent.com/assets/1892250/15055852/9a0e3352-12c3-11e6-9aca-48fc4565b52f.png)

@oxidase @TheMarex while we don't really have much routing +/-70, maybe we should wrap a conditional around this conversion for values >70 && <-70.  The poor folks at McMurdo might get totally lost.