@TheMarex 2016-05-05
It already falls back to the old code for everything outside the range.