@oxidase 2016-05-05
@danpat yes, outside the range [-70,70] the minmax error bound is no more valid and extrapolation error will be very bad. 

The full range [-90, 90] is not possible to approximate by rational polynoms because of infinities at the endpoints, so i decided that [-70,70] is a good range for approximation and for [-90,-70] and [70,90] it falls back to the old code. Another design parameter is the order of nominator/denominator, i took 11/11, because for order 12/12 double precision is not enough and the error looks like
![pade12_deg](https://cloud.githubusercontent.com/assets/4421046/15056871/38b9809c-1314-11e6-81f7-b385eb085052.png)
that looks weird for |x|>57 degrees but it is just a truncation error.