@danpat 2016-05-05
My bad for not reading properly :-P  @emiltin there's your answer :-)